Added empty redux files and moved structure of exisiting files. No redux connected yet
