+++ src/actions/actionTypes.js
@@ -1,0 +1,1 @@
+//this will hold all my action constants
