+++ src/actions/trackerAction.js
@@ -1,0 +1,2 @@
+import * as types from './actionTypes';
+//simple action that will be fired
