+++ src/reducers/trackerReducer.js
@@ -1,0 +1,1 @@
+//this will do all the object past presnt future tacker for time travel tracking
