+++ src/store/configureStore.js
@@ -1,0 +1,9 @@
+import { createStore, applyMiddleware } from 'redux';
+import rootReducer from '../reducers';
+
+export default function configureStore(initialState) {
+    return createStore(
+        rootReducer,
+        initialState,
+    );
+}
